Added infromation page to User Home page, changed main home button to send user to user homepage
Added infromation page to User Home page, changed main home button to
send user to user homepage. The idea behind this is that the if the user
is not logged in they will only see either the front page or the login
page, on these to pages we can have a different master where the main
home button will lead them back to the front page so on the rest of the
pages the user can return to their homepage with one button. My pages
are still using dummy data and are in a base stage where they are ready
to test out the database